Dont try to connect to NOT-CONFIGURED, show ram use in %

+++ main.py
@@ -134,7 +134,7 @@ try:
         else:
             SerialLog.log("Error: Unsupported Module! ", modname);
         SerialLog.log("Completed loading ", modname, " Ram Used:", ramfree - gc.mem_free())
-        SerialLog.log("Ram Allocated:", gc.mem_alloc(), "(", gc.mem_alloc()*100/(gc.mem_alloc() + gc.mem_free()), "%)")
+        SerialLog.log("Ram Allocated:", gc.mem_alloc(), "(", int(gc.mem_alloc()*100/(gc.mem_alloc() + gc.mem_free())), "%)")
 
     gc.collect()
     SerialLog.log("All code compiled. Ram Free:", gc.mem_free())
